job-analyst-app/analyst-app/job-analyst-app/analyst/requirements.txt
xlrd==1.2.0
openpyxl==3.0.9
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
et-xmlfile==1.1.0
fake-useragent==0.1.11
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Migrate==2.6.0
# flask_jwt_extended-4.4.4
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
Mako==1.2.2
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.3.5
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.41
typing-extensions==4.3.0
urllib3==1.26.8
werkzeug==2.1.2
zipp==3.8.1
flask-cors==3.0.10
pymysql==1.0.2
cryptography==38.0.1
flask_jwt_extended==4.4.4
